# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hatrick Search Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SearchResult {
  title: string;
  url: string;
  snippet: string;
  source: string;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface SearchResponse {
  query: string;
  aiAnswer: string;
  results: SearchResult[];
}

export interface SaveItemBody {
  title: string;
  url: string;
  snippet: string;
  source: string;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface SavedItem {
  id: number;
  title: string;
  url: string;
  snippet: string;
  source: string;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  savedAt: string;
}

export type SearchParams = {
  /**
   * The search query
   */
  q: string;
};