# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hatrick Search Engine API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: search
    description: Search operations
  - name: saved
    description: Saved items operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /search:
    get:
      operationId: search
      tags: [search]
      summary: Search the web
      description: Returns AI-powered answer and search result cards
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: The search query
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Missing query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /saved:
    get:
      operationId: getSavedItems
      tags: [saved]
      summary: Get all saved items
      responses:
        "200":
          description: List of saved items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedItem"

    post:
      operationId: saveItem
      tags: [saved]
      summary: Save a search result card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveItemBody"
      responses:
        "201":
          description: Item saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedItem"

  /saved/{id}:
    delete:
      operationId: deleteSavedItem
      tags: [saved]
      summary: Delete a saved item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Item deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SearchResult:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        snippet:
          type: string
        source:
          type: string
        publishedAt:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
      required:
        - title
        - url
        - snippet
        - source

    SearchResponse:
      type: object
      properties:
        query:
          type: string
        aiAnswer:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
      required:
        - query
        - aiAnswer
        - results

    SaveItemBody:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        snippet:
          type: string
        source:
          type: string
        publishedAt:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
      required:
        - title
        - url
        - snippet
        - source

    SavedItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string
        snippet:
          type: string
        source:
          type: string
        publishedAt:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        savedAt:
          type: string
      required:
        - id
        - title
        - url
        - snippet
        - source
        - savedAt