# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hatrick Search Engine API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns AI-powered answer and search result cards
 * @summary Search the web
 */
export const SearchQueryParams = zod.object({
  q: zod.coerce.string().describe("The search query"),
});

export const SearchResponse = zod.object({
  query: zod.string(),
  aiAnswer: zod.string(),
  results: zod.array(
    zod.object({
      title: zod.string(),
      url: zod.string(),
      snippet: zod.string(),
      source: zod.string(),
      publishedAt: zod.string().nullish(),
      imageUrl: zod.string().nullish(),
    }),
  ),
});

/**
 * @summary Get all saved items
 */
export const GetSavedItemsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  url: zod.string(),
  snippet: zod.string(),
  source: zod.string(),
  publishedAt: zod.string().nullish(),
  imageUrl: zod.string().nullish(),
  savedAt: zod.string(),
});
export const GetSavedItemsResponse = zod.array(GetSavedItemsResponseItem);

/**
 * @summary Save a search result card
 */
export const SaveItemBody = zod.object({
  title: zod.string(),
  url: zod.string(),
  snippet: zod.string(),
  source: zod.string(),
  publishedAt: zod.string().nullish(),
  imageUrl: zod.string().nullish(),
});

/**
 * @summary Delete a saved item
 */
export const DeleteSavedItemParams = zod.object({
  id: zod.coerce.number(),
});