# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hatrick Search Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  title: string;
  url: string;
  snippet: string;
  source: string;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}